LPS-57210 Move action and rule handlers
