+++ modules/apps/mobile-device-rules/mobile-device-rules-service/bnd.bnd
@@ -1,5 +1,8 @@
 Bundle-Name: Liferay Mobile Device Rules Service
 Bundle-SymbolicName: com.liferay.mobile.device.rules.service
 Bundle-Version: 1.0.0
-Export-Package:	com.liferay.mobile.device.rules.lar
+Export-Package:\
+	com.liferay.mobile.device.rules.lar,\
+	com.liferay.mobile.device.rules.rulegroup.action,\
+	com.liferay.mobile.device.rules.rulegroup.rule
 Include-Resource: classes
