+++ modules/apps/mobile-device-rules/mobile-device-rules-service/src/com/liferay/mobile/device/rules/lar/MDRActionStagedModelDataHandler.java
@@ -14,6 +14,7 @@
 
 package com.liferay.mobile.device.rules.lar;
 
+import com.liferay.mobile.device.rules.rulegroup.action.SiteRedirectActionHandler;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -23,7 +24,6 @@ import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xml.Element;
-import com.liferay.portal.mobile.device.rulegroup.action.impl.SiteRedirectActionHandler;
 import com.liferay.portal.model.Layout;
 import com.liferay.portal.service.LayoutLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
