+++ modules/apps/mobile-device-rules/mobile-device-rules-web/build.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE project>
 
 <project>
-	<property name="import.shared" value="mobile-device-rules-api" />
+	<property name="import.shared" value="mobile-device-rules-api,mobile-device-rules-service" />
 
 	<import file="../../../../tools/sdk/build-common-osgi-plugin.xml" />
