--- portal-impl/src/META-INF/mobile-device-spring.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-
-<beans
-	default-destroy-method="destroy"
-	default-init-method="afterPropertiesSet"
-	xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
->
-
-	<!-- Util -->
-
-	<bean class="com.liferay.portal.mobile.device.rulegroup.action.impl.SiteRedirectActionHandler" />
-	<bean class="com.liferay.portal.mobile.device.rulegroup.action.impl.SimpleRedirectActionHandler" />
-	<bean class="com.liferay.portal.mobile.device.rulegroup.action.impl.ThemeModificationActionHandler" />
-	<bean class="com.liferay.portal.mobile.device.rulegroup.action.impl.LayoutTemplateModificationActionHandler" />
-	<bean class="com.liferay.portal.mobile.device.rulegroup.rule.impl.SimpleRuleHandler" />
-</beans>
