+++ portal-impl/src/portal.properties
@@ -837,7 +837,6 @@
         META-INF/form-navigator-spring.xml,\
         META-INF/jcr-spring.xml,\
         META-INF/message-boards-spring.xml,\
-        META-INF/mobile-device-spring.xml,\
         META-INF/model-listener-spring.xml,\
         META-INF/notifications-spring.xml,\
         META-INF/poller-spring.xml,\
